# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/tags": {
      "get": {
        "operationId": "TagsAdminController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminTagEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List tags with full translation maps.",
        "tags": [
          "taxonomy"
        ]
      },
      "post": {
        "operationId": "TagsAdminController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminTagEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or slug already in use.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a tag with its translations.",
        "tags": [
          "taxonomy"
        ]
      }
    },
    "/api/v1/admin/tags/{id}": {
      "patch": {
        "operationId": "TagsAdminController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminTagEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or slug already in use.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upsert tag translations.",
        "tags": [
          "taxonomy"
        ]
      },
      "delete": {
        "operationId": "TagsAdminController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tag deleted."
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a tag.",
        "tags": [
          "taxonomy"
        ]
      }
    },
...
  },
  "info": {
    "title": "eslammuatamed API",
    "description": "REST contract for the eslammuatamed platform. Public reads are unauthenticated and locale-resolved; the /admin surface requires a bearer access token (doc 10, doc 19).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "translations[0].slug",
            "description": "Dotted path to the offending field."
          },
          "message": {
            "type": "string",
            "example": "Slug already exists for locale 'en'."
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "/problems/validation",
            "description": "A URI reference identifying the problem type."
          },
          "title": {
            "type": "string",
            "example": "Validation failed"
          },
          "status": {
            "type": "number",
            "example": 422
          },
          "detail": {
            "type": "string",
            "example": "2 fields failed validation."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/admin/articles"
          },
          "errors": {
            "description": "Present on 422 responses only.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
      "TagTranslationEntity": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "NestJS"
          },
          "slug": {
            "type": "string",
            "example": "nestjs"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "AdminTagEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "translations": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/TagTranslationEntity"
            },
            "description": "Translation map keyed by locale code."
          }
        },
        "required": [
          "id",
          "translations"
        ]
      },
      "TagTranslationDto": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "example": "en",
            "description": "Two-letter locale; must be enabled."
          },
          "name": {
            "type": "string",
            "example": "NestJS"
          },
          "slug": {
            "type": "string",
            "example": "nestjs"
          }
        },
        "required": [
          "locale",
          "name",
          "slug"
        ]
      },
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "At least one locale translation.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagTranslationDto"
            }
          }
        },
        "required": [
          "translations"
        ]
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagTranslationDto"
            }
          }
        }
      },
...
    }
  }
}